🚚 rename : RegionPopupContent로 이름 변경
컴포넌트 성격에 따른 이름 변경 완료
