+++ src/components/molecules/index.ts
@@ -1,6 +1,6 @@
 export { default as Molecule } from './Molecule'
 export { default as GenrePopupContent } from './GenrePopupContent'
-export { default as RegionButtonGroup } from './RegionButtonGroup'
+export { default as RegionPopupContent } from './RegionPopupContent'
 export { default as StatusPopupContent } from './StatusPopupContent'
 export { default as ToggleText } from './ToggleText'
 export { default as FormCategorySelect } from './FormCategorySelect'
